+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added `formatDate` function to `PdfDocument` class.
 - Removed `composer.lock` from export.
 
 ## 2.0.13 - 2025-01-15